+++ app/Http/Controllers/CountryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Helpers\JavaScriptHelper;
 use App\Iframes\CityIframe;
 use App\Iframes\CountryIframe;
 use App\Models\Country;
@@ -70,6 +71,10 @@ class CountryController extends Controller
 
     public function destroy(Country $country)
     {
+        if ($country->residences()->count() > 0) {
+            return JavaScriptHelper::alert("You can't delete $country->name country!");
+        }
+
         $country->cities()->delete();
         $country->delete();
